fix: refactored the fest-route to remove the need for conditional rendering of fest-discussion. The two routes are now independent of each other and have their logic separated. This allowed for fest-route to be cleaner
