+++ src/graphql/get-discussions.js
@@ -3,6 +3,7 @@ import { gql } from "@apollo/client";
 export const GET_DISCUSSIONS = gql`
   query Fest($where: FestWhereUniqueInput!) {
     fest(where: $where) {
+      id
       festNotes {
         content
         createdAt
